Revert "Transition to async mysql driver"
This reverts commit f60e27b326d9e7260a6e02317bc256d7daedf5d1.

+++ app/routers/ideas.py
@@ -1,6 +1,5 @@
 from fastapi import APIRouter, Depends
-import asyncio
-import aiomysql
+import mysql.connector
 from datetime import datetime
 from typing import Optional
 
@@ -17,13 +16,13 @@ router = APIRouter(
     tags=["ideas"]
 )
 
-loop = asyncio.get_event_loop()
+db = mysql.connector.connect()
 
 
-async def is_db_up():
+def is_db_up():
     try:
-        await db.ping(reconnect=True)
-    except aiomysql.InterfaceError as ex:
+        db.ping(reconnect=True, attempts=3, delay=5)
+    except mysql.connector.errors.InterfaceError as ex:
         raise HTTPException(status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail=ex.__dict__)
     return True
 
@@ -31,16 +30,12 @@ async def is_db_up():
 # Event that set-ups the application for startup
 @router.on_event("startup")
 async def startup_event():
-    # noinspection PyGlobalUndefined
     global db
-    db = await aiomysql.connect(
+    db = mysql.connector.connect(
         host=DB_HOST,
-        port=3306,
         user=DB_USER,
         password=DB_PASS,
-        db=DB_NAME,
-        autocommit=True,
-        loop=loop
+        database=DB_NAME
     )
     # This makes it work without having to commit after every query
     db.autocommit = True
@@ -48,8 +43,9 @@ async def startup_event():
 
 @router.get("/get", response_model=IdeasList)
 async def get_ideas(page: Optional[int] = 0, cat: Optional[str] = None):
-    await is_db_up()
-    cursor = await db.cursor(aiomysql.DictCursor)
+    is_db_up()
+
+    cursor = db.cursor(dictionary=True)
     query = "SELECT " \
             "ideas.id, seller_id, title, short_desc, date_publish, date_expiry, price, " \
             "files.public_path AS image_url," \
@@ -58,31 +54,35 @@ async def get_ideas(page: Optional[int] = 0, cat: Optional[str] = None):
             "WHERE buyer_id IS NULL AND " \
             "(%s IS NULL OR ideas.id IN (SELECT idea_id FROM ideas_categories WHERE category LIKE %s))" \
             "ORDER BY date_publish DESC LIMIT %s, %s "
-    await cursor.execute(query, (cat, cat, page * 10, (page + 1) * 10))
-    results = await cursor.fetchall()
+    cursor.execute(
+        query, (cat, cat, page * 10, (page + 1) * 10)
+    )
+    results = cursor.fetchall()
+
     # Check if there are results, don't waste time to continue if not
     if len(results) == 0:
         return IdeasList(countLeft=0, ideas=list())
 
     # Get categories
     for result in results:
-        await cursor.execute("SELECT category FROM ideas_categories WHERE idea_id=%s", (result["id"],))
-        result["categories"] = list(map(lambda x: x["category"], await cursor.fetchall()))
+        cursor.execute("SELECT category FROM ideas_categories WHERE idea_id=%s", (result["id"],))
+        result["categories"] = list(map(lambda x: x["category"], cursor.fetchall()))
 
     # Find the number of ideas matching the criteria
     query = "SELECT COUNT(*) AS ideas_count " \
             "FROM ideas " \
             "WHERE ideas.buyer_id IS NULL AND " \
             "(%s IS NULL OR ideas.id IN (SELECT idea_id FROM ideas_categories WHERE category LIKE %s))"
-    await cursor.execute(query, (cat, cat))
-    ideas_count = await cursor.fetchone()
+    cursor.execute(query, (cat, cat))
+    ideas_count = cursor.fetchone()["ideas_count"]
+
     # Calculate remaining ideas for endless scrolling feature
     if page == 0:
-        ideas_left = ideas_count["ideas_count"] - len(results)
+        ideas_left = ideas_count - len(results)
     else:
-        ideas_left = ideas_count["ideas_count"] - (page * 10 + len(results))
+        ideas_left = ideas_count - (page * 10 + len(results))
 
-    await cursor.close()
+    cursor.close()
 
     return IdeasList(
         countLeft=ideas_left,
@@ -104,16 +104,16 @@ async def get_ideas(page: Optional[int] = 0, cat: Optional[str] = None):
 @router.get("/get/{idea_id}", response_model=IdeaFull)
 async def get_idea_by_id(idea_id: str, token_data: AccessToken = Depends(get_token_data)):
     verify_idea_id(idea_id)
-    await is_db_up()
+    is_db_up()
 
-    cursor = await db.cursor(aiomysql.DictCursor)
+    cursor = db.cursor(dictionary=True)
     query = "SELECT ideas.*, " \
             "(SELECT COUNT(*) FROM ideas_likes WHERE idea_id=ideas.id) AS likes, " \
             "files.public_path AS image_url " \
             "FROM ideas LEFT JOIN files ON ideas.id=files.id " \
             "WHERE ideas.id = %s"
-    await cursor.execute(query, (idea_id,))
-    result = await cursor.fetchone()
+    cursor.execute(query, (idea_id,))
+    result = cursor.fetchone()
 
     if result is None:
         raise IdeaNotFoundError
@@ -121,17 +121,17 @@ async def get_idea_by_id(idea_id: str, token_data: AccessToken = Depends(get_tok
         raise IdeaAccessDeniedError
 
     # Get categories
-    await cursor.execute("SELECT category FROM ideas_categories WHERE idea_id=%s", (result["id"],))
-    result["categories"] = list(map(lambda x: x["category"], await cursor.fetchall()))
+    cursor.execute("SELECT category FROM ideas_categories WHERE idea_id=%s", (result["id"],))
+    result["categories"] = list(map(lambda x: x["category"], cursor.fetchall()))
 
     # Check if the user is the owner of the idea, if so fetch the files, else remove long description
     if result["buyer_id"] != token_data.user_id:
         del result["long_desc"]
     else:
-        await cursor.execute("SELECT * FROM files WHERE idea_id=%s AND idea_id!=id", (result["id"],))
-        result["files"] = list(await cursor.fetchall())
+        cursor.execute("SELECT * FROM files WHERE idea_id=%s AND idea_id!=id", (result["id"],))
+        result["files"] = list(cursor.fetchall())
 
-    await cursor.close()
+    cursor.close()
 
     return IdeaFull(
         id=result["id"],
@@ -161,17 +161,18 @@ async def get_idea_by_id(idea_id: str, token_data: AccessToken = Depends(get_tok
 
 
 @router.get("/get-hottest", response_model=IdeasHottest)
-async def get_hottest_ideas():
-    await is_db_up()
+def get_hottest_ideas():
+    is_db_up()
 
-    cursor = await db.cursor(aiomysql.DictCursor)
+    cursor = db.cursor(dictionary=True)
     query = "SELECT ideas.id, ideas.title, files.public_path AS image_url, " \
             "(SELECT COUNT(*) FROM ideas_likes WHERE idea_id=ideas.id) AS likes " \
             "FROM ideas LEFT JOIN files ON ideas.id=files.id " \
             "WHERE buyer_id IS NULL ORDER BY likes DESC LIMIT 5"
-    await cursor.execute(query)
-    results = await cursor.fetchall()
-    await cursor.close()
+    cursor.execute(query)
+    results = cursor.fetchall()
+
+    cursor.close()
 
     return IdeasHottest(
         ideas=list(map(lambda idea: IdeaSmall(
@@ -185,28 +186,28 @@ async def get_hottest_ideas():
 
 @router.post("/post")
 async def post_idea(idea: IdeaPost, token_data: AccessToken = Depends(get_token_data)):
-    await is_db_up()
+    is_db_up()
 
     # Long description is used for id of the idea, because it needs to be unique
     idea_id = calculate_idea_id(idea.long_desc)
 
-    cursor = await db.cursor(aiomysql.DictCursor)
+    cursor = db.cursor()
     query = "INSERT INTO ideas(id, seller_id, title, short_desc, long_desc, " \
             "date_publish, date_expiry, price) VALUES(%s, %s, %s, %s, %s, %s, %s, %s)"
     data = (idea_id, token_data.user_id, idea.title, idea.short_desc, idea.long_desc, datetime.now().isoformat(),
             (datetime.now() + IDEA_EXPIRES_AFTER).isoformat(), idea.price)
     try:
-        await cursor.execute(query, data)
+        cursor.execute(query, data)
         for category in idea.categories:
-            await cursor.execute("INSERT INTO ideas_categories(idea_id, category) VALUES(%s, %s)", (idea_id, category))
-    except aiomysql.IntegrityError as ex:
+            cursor.execute("INSERT INTO ideas_categories(idea_id, category) VALUES(%s, %s)", (idea_id, category))
+    except mysql.connector.errors.IntegrityError as ex:
         field = ex.msg.split()[5]
         if field == "'id'":
             raise IdeaDuplicationError
         else:
             raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=ex.__dict__)
 
-    await cursor.close()
+    cursor.close()
 
     return idea_id
 
@@ -214,18 +215,15 @@ async def post_idea(idea: IdeaPost, token_data: AccessToken = Depends(get_token_
 @router.put("/like", response_model=Like)
 async def like_idea(idea_id: str, token_data: AccessToken = Depends(get_token_data)):
     verify_idea_id(idea_id)
-    await is_db_up()
+    is_db_up()
 
-    cursor = await db.cursor(aiomysql.DictCursor)
+    cursor = db.cursor(dictionary=True)
     # Try to insert a like row in the table, if a duplication error is thrown, delete the like
     try:
-        await cursor.execute("INSERT INTO ideas_likes(idea_id, user_id) VALUES(%s, %s)", (idea_id, token_data.user_id))
+        cursor.execute("INSERT INTO ideas_likes(idea_id, user_id) VALUES(%s, %s)", (idea_id, token_data.user_id))
         is_liked = True
-    except aiomysql.IntegrityError:
-        await cursor.execute(
-            "DELETE FROM ideas_likes WHERE idea_id = %s AND user_id = %s",
-            (idea_id, token_data.user_id)
-        )
+    except mysql.connector.IntegrityError:
+        cursor.execute("DELETE FROM ideas_likes WHERE idea_id = %s AND user_id = %s", (idea_id, token_data.user_id))
         is_liked = False
 
     # Get the number of likes
@@ -233,9 +231,9 @@ async def like_idea(idea_id: str, token_data: AccessToken = Depends(get_token_da
             "COUNT(*) AS likes_count, " \
             "((SELECT COUNT(*) FROM ideas_likes WHERE idea_id=%s AND user_id=%s) = 1) AS is_liked " \
             "FROM ideas_likes WHERE idea_id = %s"
-    await cursor.execute(query, (idea_id, token_data.user_id, idea_id))
-    result = await cursor.fetchone()
-    await cursor.close()
+    cursor.execute(query, (idea_id, token_data.user_id, idea_id))
+    result = cursor.fetchone()
+    cursor.close()
 
     if result is None:
         raise IdeaNotFoundError
@@ -248,4 +246,4 @@ async def like_idea(idea_id: str, token_data: AccessToken = Depends(get_token_da
 
 @router.on_event("shutdown")
 async def shutdown_event():
-    await db.close()
+    db.close()
